Moved Camera logic out from navigation folder to components/Camera
