+++ src/components/Camera/ScannerPageStyleSheet.ts
