+++ src/navigation/tabs/Camera/index.tsx
@@ -1,6 +1,6 @@
 import { createStackNavigator } from "react-navigation";
-import { ScannerPage } from "./ScannerPage";
-import ProfilePreviewPage from "./ProfilePreviewPage";
+import { ScannerPage } from "../../../components/Camera/ScannerPage";
+import ProfilePreviewPage from "../../../components/Camera/ProfilePreviewPage";
 
 const CameraTab = createStackNavigator(
   {
